Post_Service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
protobuf==4.25.1
confluent-kafka==2.3.0
